+++ lib/ui-desktop-primary.sh
@@ -137,7 +137,7 @@ ui_draw_desktop() {
 
     ui_box_rule "$width" "$UI_TL" "$UI_TR"
     ui_box_center_line "$width" "$title" title
-    ui_desktop_header_rule "$width" 'AHORA SUENA' "FAVORITOS (${#FAVORITE_NAMES[@]})"
+    ui_desktop_header_rule "$width" 'AHORA SUENA' "EMISORAS FAVORITAS (${#FAVORITE_NAMES[@]})"
 
     local station station_style marker favorite_badge recording_badge state_badge
     marker=$(ui_player_marker)